Hi
Hello and welcome to
My Photography Portfolio
James Carmelo is a 16 year old photographer
who began my journey at the age of 16, driven
by an insatiable curiosity and a desire to
capture the world in unique and compelling
ways. Photography is more than a skill—it’s an
art form that allows me to transform fleeting
moments into timeless stories. With every
shot, he aims to capture the essence of my
subjects and to convey emotions that
resonate deeply with the clients.